Completed

+++ app/Http/Controllers/JobController.php
@@ -3,64 +3,75 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
-use Illuminate\Http\Request;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
     public function index()
     {
+        $jobs = Job::with('employer')->latest()->simplePaginate(3);
 
-        $jobs = Job::with('employer')->latest()->paginate(3);
         return view('jobs.index', [
             'jobs' => $jobs
         ]);
     }
+
     public function create()
     {
         return view('jobs.create');
-
     }
+
+    public function show(Job $job)
+    {
+        return view('jobs.show', ['job' => $job]);
+    }
+
     public function store()
     {
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
+
         Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
+
         return redirect('/jobs');
     }
+
+    public function edit(Job $job)
+    {
+        return view('jobs.edit', ['job' => $job]);
+    }
+
     public function update(Job $job)
     {
+        Gate::authorize('edit', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
+
         $job->update([
             'title' => request('title'),
             'salary' => request('salary'),
         ]);
+
         return redirect('/jobs/' . $job->id);
     }
+
     public function destroy(Job $job)
     {
+        Gate::authorize('edit', $job);
+
         $job->delete();
+
         return redirect('/jobs');
     }
-    public function show(Job $job)
-    {
-        return view('jobs.show', [
-            'jobs' => $job
-        ]);
-    }
-    public function edit(Job $job)
-    {
-        return view('jobs.edit', [
-            'jobs' => $job
-        ]);
-    }
 }
-
